requirements.txt
absl-py==2.1.0
arabic-reshaper==3.0.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
deep-translator==1.11.4
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mediapipe==0.10.9
ml-dtypes==0.3.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
requests==2.31.0
scipy==1.12.0
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
sympy==1.12
torch==2.2.1
typing_extensions==4.10.0
urllib3==2.2.1
zipp==3.18.1
